+++ java/spark-1.5.2/invoke.scala
@@ -16,8 +16,10 @@ class Invoke {
   // Returns an Option[Int] which is the index of the matched signature in the array.
   def findMatchedSignature(
     parameterTypesOfMethods: Array[Array[Class[_]]],
-    args: Array[Object]): Option[Int] = {
+    args: Array[Object],
+    logger: Logger): Option[Int] = {
       val numArgs = args.length
+      var mismatches = Array("")
 
       for (index <- 0 until parameterTypesOfMethods.length) {
         val parameterTypes = parameterTypesOfMethods(index)
@@ -62,6 +64,8 @@ class Invoke {
               if ((parameterType.isPrimitive || args(i) != null) &&
                   !parameterWrapperType.isInstance(args(i))) {
                 argMatched = false
+
+                mismatches = mismatches ++ Array("method instance " + index + " with " + parameterTypes.length + " parameters for parameter " + i)
               }
             }
 
@@ -108,6 +112,9 @@ class Invoke {
           }
         }
       }
+
+      logger.logWarning(s"Failed to match" + mismatches.mkString(", ") + ".")
+
       None
     }
 
@@ -124,7 +131,8 @@ class Invoke {
     if (selectedMethods.length > 0) {
       val index = findMatchedSignature(
         selectedMethods.map(_.getParameterTypes),
-        args)
+        args,
+        logger)
 
       if (index.isEmpty) {
         logger.logWarning(
@@ -142,7 +150,8 @@ class Invoke {
       val ctors = cls.getConstructors
       val index = findMatchedSignature(
         ctors.map(_.getParameterTypes),
-        args)
+        args,
+        logger)
 
       if (index.isEmpty) {
         logger.logWarning(